+++ src/xcore/keymap.cpp
@@ -456,11 +456,21 @@ static xShortcut short_tab[] = {
 	{SCG_DEBUGA, XCUT_LABLIST, "key.dbg.lablist", "DeBUGa: Show labels list", QKeySequence(), QKeySequence()},
 	{SCG_DEBUGA, XCUT_DBG_RELOAD, "key.dbg.reload", "DeBUGa: Reload snapshot and labels", QKeySequence(), QKeySequence()},
 
-	{SCG_DISASM, XCUT_TOPC, "key.disasm.topc", "Diasm: Jump to PC", QKeySequence(), QKeySequence(Qt::Key_Home)},
-	{SCG_DISASM, XCUT_SETPC, "key.disasm.setpc", "Diasm: Set PC", QKeySequence(), QKeySequence(Qt::Key_End)},
-	{SCG_DISASM, XCUT_SETBRK, "key.disasm.setbrk", "Diasm: Breakpoint", QKeySequence(), QKeySequence(Qt::Key_Space)},
-	{SCG_DISASM, XCUT_JUMPTO, "key.disasm.jump", "Diasm: Jump to operand", QKeySequence(), QKeySequence(Qt::Key_F4)},
-	{SCG_DISASM, XCUT_RETFROM, "key.disasm.ret", "Diasm: Return", QKeySequence(), QKeySequence(Qt::Key_F5)},
+	{SCG_DISASM, XCUT_TOPC, "key.disasm.topc", "Disasm: Jump to PC", QKeySequence(), QKeySequence(Qt::Key_Home)},
+	{SCG_DISASM, XCUT_SETPC, "key.disasm.setpc", "Disasm: Set PC", QKeySequence(), QKeySequence(Qt::Key_End)},
+	{SCG_DISASM, XCUT_SETBRK, "key.disasm.setbrk", "Disasm: Breakpoint", QKeySequence(), QKeySequence(Qt::Key_Space)},
+	{SCG_DISASM, XCUT_JUMPTO, "key.disasm.jump", "Disasm: Jump to operand", QKeySequence(), QKeySequence(Qt::Key_F4)},
+	{SCG_DISASM, XCUT_RETFROM, "key.disasm.ret", "Disasm: Return", QKeySequence(), QKeySequence(Qt::Key_F5)},
+	{SCG_DISASM, XCUT_GOTOADR, "key.disasm.goto", "Disasm: Go to address", QKeySequence(), QKeySequence(Qt::Key_G)},
+
+	{SCG_DUMP, XCUT_DUMP_GOTOADR, "key.dump.goto", "Dump: Go to address", QKeySequence(), QKeySequence(Qt::Key_G)},
+	{SCG_DUMP, XCUT_DUMP_REG_PC, "key.dump.goto.pc", "Dump: Go to (PC)", QKeySequence(), QKeySequence(Qt::CTRL | Qt::Key_P)},
+	{SCG_DUMP, XCUT_DUMP_REG_SP, "key.dump.goto.sp", "Dump: Go to (SP)", QKeySequence(), QKeySequence(Qt::CTRL | Qt::Key_S)},
+	{SCG_DUMP, XCUT_DUMP_REG_BC, "key.dump.goto.bc", "Dump: Go to (BC)", QKeySequence(), QKeySequence(Qt::CTRL | Qt::Key_B)},
+	{SCG_DUMP, XCUT_DUMP_REG_DE, "key.dump.goto.de", "Dump: Go to (DE)", QKeySequence(), QKeySequence(Qt::CTRL | Qt::Key_D)},
+	{SCG_DUMP, XCUT_DUMP_REG_HL, "key.dump.goto.hl", "Dump: Go to (HL)", QKeySequence(), QKeySequence(Qt::CTRL | Qt::Key_H)},
+	{SCG_DUMP, XCUT_DUMP_REG_IX, "key.dump.goto.ix", "Dump: Go to (IX)", QKeySequence(), QKeySequence(Qt::CTRL | Qt::Key_X)},
+	{SCG_DUMP, XCUT_DUMP_REG_IY, "key.dump.goto.iy", "Dump: Go to (IY)", QKeySequence(), QKeySequence(Qt::CTRL | Qt::Key_Y)},
 
 	{0, -1, NULL, NULL, QKeySequence(), QKeySequence()}
 };